+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-apollo",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "React data container for Apollo Client",
   "main": "lib/react-apollo.umd.js",
   "module": "./lib/index.js",
@@ -133,7 +133,7 @@
     "uglify-js": "^2.6.2"
   },
   "dependencies": {
-    "apollo-client": "^1.2.1",
+    "apollo-client": "^1.2.2",
     "graphql-anywhere": "^3.0.0",
     "graphql-tag": "^2.0.0",
     "hoist-non-react-statics": "^1.2.0",
