+++ src/browser.ts
@@ -1,5 +1,6 @@
 export { default as ApolloProvider } from './ApolloProvider';
-export { default as graphql, withApollo, InjectedGraphQLProps } from './graphql';
+export { default as graphql, InjectedGraphQLProps } from './graphql';
+export { withApollo } from './withApollo';
 
 // expose easy way to join queries from redux
 export { compose } from 'redux';
